fix(core): improve `DictWordMetadata::difference`
Generate the `difference` function using a macro, to ensure no relevant
functions used for calculating the difference are missed.

Also use a larger integer type for the return value to reduce the chance
of accidental overflows in the future.

+++ harper-core/src/thesaurus_helper.rs
@@ -37,7 +37,7 @@ pub fn get_synonyms_sorted(_word: &str, _token: &TokenKind) -> Option<Vec<&'stat
                 if let Some(syn_meta) = dict.get_word_metadata_str(syn) {
                     word_meta.difference(&syn_meta)
                 } else {
-                    u8::MAX
+                    u32::MAX
                 }
             });
         }
